Add generic methods in APIClient interface
