+++ cli/client/interface.go
@@ -32,6 +32,15 @@ type APIClient interface {
 
 // GenericClient exposes generic resource methods.
 type GenericClient interface {
+	// Delete deletes the key with the given path
+	Delete(path string) error
+	// Get retrieves the key at the given path and stores it into obj
+	Get(path string, obj interface{}) error
+	// List retrieves all keys with the given path prefix and stores them into objs
+	List(path string, objs interface{}) error
+	// Post creates the given obj at the specified path
+	Post(path string, obj interface{}) error
+
 	// PutResource puts a resource according to its URIPath.
 	PutResource(types.Resource) error
 }
